Tableau, Python, Splinter, Urllib, Pandas
This is a simple project illustrating the how multiple modalities can combine to grab, clean and visualize data.  In this case, I used information readily available from NY State for the collaboration with Citibank to provide bike rentals throughout the Greater Manhattan area. The analysis itself is not particularly complicated. With more time, I would have included a larger database in order to do forecasting.